Fix(?) SIP plan ordering [..]
This  is a  very brittle  solution, but  the best  I
could think of given the  poor legacy DB design. The
gist is  that if  the name  of the  SIP plan  is not
along the lines of

    11/27/2023 - In-home - Inst Ructor
    11/27/23 - Support Group - Faci Litator

then it  will crash as  the Django query  breaks the
string of the  plan name at the first  `-` (seems to
be  non-greedy) and  the substring  until the  first
hyphen has to be `Cast`-able to `DateField`.

Fortunately, `Cast`  doesn't seem to squeamish  as I
have stumbled  upon a couple of  instances where the
site does not crash, but I thought it would. Fingers
crossed.