[bulk notes] Optimize editing service events [..]
The static way meant that every client `<select>` options started to load
about a 1000 client names and ids, so the page load time was around 5 sec
for 30 clients per service event.

Pushed it down to 0.5s with HTMX and some `fetch`; not sure why the HTMX
attributes didn't work for the existing clients, but not going to sweat
it.