[bulk notes] Fix(?) client dropdown(s) [..]
The client dropdowns on `oib_service_event_add` didn't play well
with JAWS and also kept crashing the page is someone clicked on
the dropdown quickly multiple times.

It looks like the root cause for both was the same: I set `hx-get`
to **always** get the client list on mousedown, instead of getting
it only once (on focus).

(?) - This still needs to be tested with JAWS